# add more inlines for better efficiency