--- a/k256/benches/field.rs
+++ b/k256/benches/field.rs
@@ -1,7 +1,7 @@
 //! secp256k1 field element benchmarks
 
 use criterion::{
-    criterion_group, criterion_main, measurement::Measurement, BenchmarkGroup, Criterion,
+    black_box, criterion_group, criterion_main, measurement::Measurement, BenchmarkGroup, Criterion,
 };
 use k256::FieldElement;
 
@@ -31,33 +31,35 @@ fn test_field_element_y() -> FieldElement {
 
 fn bench_field_element_normalize_weak<'a, M: Measurement>(group: &mut BenchmarkGroup<'a, M>) {
     let x = test_field_element_x();
-    group.bench_function("normalize_weak", |b| b.iter(|| x.normalize_weak()));
+    group.bench_function("normalize_weak", |b| {
+        b.iter(|| black_box(x).normalize_weak())
+    });
 }
 
 fn bench_field_element_normalize<'a, M: Measurement>(group: &mut BenchmarkGroup<'a, M>) {
     let x = test_field_element_x();
-    group.bench_function("normalize", |b| b.iter(|| x.normalize()));
+    group.bench_function("normalize", |b| b.iter(|| black_box(x).normalize()));
 }
 
 fn bench_field_element_mul<'a, M: Measurement>(group: &mut BenchmarkGroup<'a, M>) {
     let x = test_field_element_x();
     let y = test_field_element_y();
-    group.bench_function("mul", |b| b.iter(|| &x * &y));
+    group.bench_function("mul", |b| b.iter(|| &black_box(x) * &black_box(y)));
 }
 
 fn bench_field_element_square<'a, M: Measurement>(group: &mut BenchmarkGroup<'a, M>) {
     let x = test_field_element_x();
-    group.bench_function("square", |b| b.iter(|| x.square()));
+    group.bench_function("square", |b| b.iter(|| black_box(x).square()));
 }
 
 fn bench_field_element_sqrt<'a, M: Measurement>(group: &mut BenchmarkGroup<'a, M>) {
     let x = test_field_element_x();
-    group.bench_function("sqrt", |b| b.iter(|| x.sqrt()));
+    group.bench_function("sqrt", |b| b.iter(|| black_box(x).sqrt()));
 }
 
 fn bench_field_element_invert<'a, M: Measurement>(group: &mut BenchmarkGroup<'a, M>) {
     let x = test_field_element_x();
-    group.bench_function("invert", |b| b.iter(|| x.invert()));
+    group.bench_function("invert", |b| b.iter(|| black_box(x).invert()));
 }
 
 fn bench_field_element(c: &mut Criterion) {