--- a/k256/src/arithmetic/affine.rs
+++ b/k256/src/arithmetic/affine.rs
@@ -120,6 +120,7 @@ impl AffineCoordinates for AffinePoint {
 }
 
 impl ConditionallySelectable for AffinePoint {
+    #[inline(always)]
     fn conditional_select(a: &AffinePoint, b: &AffinePoint, choice: Choice) -> AffinePoint {
         AffinePoint {
             x: FieldElement::conditional_select(&a.x, &b.x, choice),
@@ -130,6 +131,7 @@ impl ConditionallySelectable for AffinePoint {
 }
 
 impl ConstantTimeEq for AffinePoint {
+    #[inline(always)]
     fn ct_eq(&self, other: &AffinePoint) -> Choice {
         (self.x.negate(1) + &other.x).normalizes_to_zero()
             & (self.y.negate(1) + &other.y).normalizes_to_zero()
@@ -146,6 +148,7 @@ impl Default for AffinePoint {
 impl DefaultIsZeroes for AffinePoint {}
 
 impl PartialEq for AffinePoint {
+    #[inline(always)]
     fn eq(&self, other: &AffinePoint) -> bool {
         self.ct_eq(other).into()
     }
@@ -156,6 +159,7 @@ impl Eq for AffinePoint {}
 impl Mul<Scalar> for AffinePoint {
     type Output = ProjectivePoint;
 
+    #[inline(always)]
     fn mul(self, scalar: Scalar) -> ProjectivePoint {
         ProjectivePoint::from(self) * scalar
     }
@@ -164,6 +168,7 @@ impl Mul<Scalar> for AffinePoint {
 impl Mul<&Scalar> for AffinePoint {
     type Output = ProjectivePoint;
 
+    #[inline(always)]
     fn mul(self, scalar: &Scalar) -> ProjectivePoint {
         ProjectivePoint::from(self) * scalar
     }
@@ -172,6 +177,7 @@ impl Mul<&Scalar> for AffinePoint {
 impl Neg for AffinePoint {
     type Output = AffinePoint;
 
+    #[inline(always)]
     fn neg(self) -> Self::Output {
         AffinePoint {
             x: self.x,
@@ -182,6 +188,7 @@ impl Neg for AffinePoint {
 }
 
 impl DecompressPoint<Secp256k1> for AffinePoint {
+    #[inline(always)]
     fn decompress(x_bytes: &FieldBytes, y_is_odd: Choice) -> CtOption<Self> {
         FieldElement::from_bytes(x_bytes).and_then(|x| {
             let alpha = (x * &x * &x) + &CURVE_EQUATION_B;
@@ -205,6 +212,7 @@ impl DecompressPoint<Secp256k1> for AffinePoint {
 ///
 /// [BIP 340]: https://github.com/bitcoin/bips/blob/master/bip-0340.mediawiki
 impl DecompactPoint<Secp256k1> for AffinePoint {
+    #[inline(always)]
     fn decompact(x_bytes: &FieldBytes) -> CtOption<Self> {
         Self::decompress(x_bytes, Choice::from(0))
     }
@@ -213,6 +221,7 @@ impl DecompactPoint<Secp256k1> for AffinePoint {
 impl GroupEncoding for AffinePoint {
     type Repr = CompressedPoint;
 
+    #[inline(always)]
     fn from_bytes(bytes: &Self::Repr) -> CtOption<Self> {
         EncodedPoint::from_bytes(bytes)
             .map(|point| CtOption::new(point, Choice::from(1)))
@@ -225,11 +234,13 @@ impl GroupEncoding for AffinePoint {
             .and_then(|point| Self::from_encoded_point(&point))
     }
 
+    #[inline(always)]
     fn from_bytes_unchecked(bytes: &Self::Repr) -> CtOption<Self> {
         // No unchecked conversion possible for compressed points
         Self::from_bytes(bytes)
     }
 
+    #[inline(always)]
     fn to_bytes(&self) -> Self::Repr {
         let encoded = self.to_encoded_point(true);
         let mut result = CompressedPoint::default();
@@ -244,6 +255,7 @@ impl FromEncodedPoint<Secp256k1> for AffinePoint {
     /// # Returns
     ///
     /// `None` value if `encoded_point` is not on the secp256k1 curve.
+    #[inline(always)]
     fn from_encoded_point(encoded_point: &EncodedPoint) -> CtOption<Self> {
         match encoded_point.coordinates() {
             sec1::Coordinates::Identity => CtOption::new(Self::IDENTITY, 1.into()),
@@ -270,6 +282,7 @@ impl FromEncodedPoint<Secp256k1> for AffinePoint {
 }
 
 impl ToEncodedPoint<Secp256k1> for AffinePoint {
+    #[inline(always)]
     fn to_encoded_point(&self, compress: bool) -> EncodedPoint {
         EncodedPoint::conditional_select(
             &EncodedPoint::from_affine_coordinates(