If I add more `inline`s, the performance in most current benchmarks improves on my machine (AMD EPYC 7302). Benchmarks were run as `RUSTFLAGS='-C target-cpu=native' cargo bench --features expose-field`.

```
    Running benches/ecdsa.rs
ecdsa/try_sign_prehashed
                        time:   [48.866 µs 48.901 µs 48.944 µs]
                        change: [-10.073% -8.5345% -6.8744%] (p = 0.00 < 0.05)
ecdsa/verify_prehashed  
                        time:   [103.95 µs 104.43 µs 105.00 µs]
                        change: [-12.366% -11.807% -11.312%] (p = 0.00 < 0.05)

     Running benches/field.rs
field element operations/normalize_weak
                        time:   [3.1100 ns 3.1228 ns 3.1391 ns]
                        change: [-44.824% -44.193% -43.486%] (p = 0.00 < 0.05)
field element operations/normalize
                        time:   [10.600 ns 10.665 ns 10.757 ns]
                        change: [-25.955% -25.278% -24.669%] (p = 0.00 < 0.05)
field element operations/mul
                        time:   [22.809 ns 22.943 ns 23.120 ns]
                        change: [-22.698% -22.031% -21.383%] (p = 0.00 < 0.05)
field element operations/square
                        time:   [17.109 ns 17.158 ns 17.223 ns]
                        change: [-24.579% -24.168% -23.815%] (p = 0.00 < 0.05)
field element operations/invert
                        time:   [4.8184 µs 4.8271 µs 4.8384 µs]
                        change: [-26.792% -26.425% -26.113%] (p = 0.00 < 0.05)
field element operations/sqrt
                        time:   [4.9200 µs 4.9255 µs 4.9332 µs]
                        change: [-24.270% -23.922% -23.657%] (p = 0.00 < 0.05)

     Running benches/scalar.rs
high-level operations/point-scalar mul
                        time:   [56.610 µs 56.936 µs 57.360 µs]
                        change: [-11.544% -10.793% -10.106%] (p = 0.00 < 0.05)
high-level operations/mul_by_generator naive
                        time:   [56.326 µs 56.422 µs 56.535 µs]
                        change: [-11.644% -11.084% -10.507%] (p = 0.00 < 0.05)
high-level operations/mul_by_generator precomputed
                        time:   [27.987 µs 28.020 µs 28.066 µs]
                        change: [-9.8668% -8.9146% -8.1681%] (p = 0.00 < 0.05)
high-level operations/lincomb via mul+add
                        time:   [112.67 µs 113.36 µs 114.37 µs]
                        change: [-9.3266% -8.7609% -8.0640%] (p = 0.00 < 0.05)
high-level operations/lincomb()
                        time:   [89.727 µs 90.030 µs 90.401 µs]
                        change: [-9.4576% -8.5924% -7.8728%] (p = 0.00 < 0.05)
scalar operations/sub   
                        time:   [6.2334 ns 6.2527 ns 6.2782 ns]
                        change: [-18.510% -17.243% -16.189%] (p = 0.00 < 0.05)
scalar operations/add   
                        time:   [8.1355 ns 8.1517 ns 8.1712 ns]
                        change: [-20.691% -19.968% -19.352%] (p = 0.00 < 0.05)
scalar operations/mul   
                        time:   [46.323 ns 46.443 ns 46.596 ns]
                        change: [-0.6997% -0.1909% +0.2646%] (p = 0.46 > 0.05)
scalar operations/negate
                        time:   [7.6674 ns 7.6991 ns 7.7444 ns]
                        change: [-19.449% -19.121% -18.801%] (p = 0.00 < 0.05)
scalar operations/invert
                        time:   [15.156 µs 15.167 µs 15.180 µs]
                        change: [-1.0203% -0.4949% -0.1341%] (p = 0.01 < 0.05)
```

This PR also adds `criterion::black_box()` to the input variables in benchmarks s.t. they are not optimized away.