@altkdf 2023-12-05
On my Mac (M1), the numbers look similar. It would be nice to benchmark this also on a simple x86 machine. 
```
     Running benches/ecdsa.rs
ecdsa/try_sign_prehashed
                        time:   [33.505 µs 33.559 µs 33.626 µs]
                        change: [-9.5527% -9.4414% -9.3221%] (p = 0.00 < 0.05)
ecdsa/verify_prehashed  
                        time:   [64.628 µs 64.705 µs 64.825 µs]
                        change: [-12.094% -11.987% -11.868%] (p = 0.00 < 0.05)

     Running benches/field.rs
field element operations/normalize_weak
                        time:   [2.0719 ns 2.0731 ns 2.0749 ns]
                        change: [-66.954% -65.750% -64.414%] (p = 0.00 < 0.05)
field element operations/normalize
                        time:   [10.731 ns 10.733 ns 10.736 ns]
                        change: [-4.0784% -3.0586% -1.9188%] (p = 0.00 < 0.05)
field element operations/mul
                        time:   [14.358 ns 14.364 ns 14.371 ns]
                        change: [-13.244% -13.133% -13.035%] (p = 0.00 < 0.05)
field element operations/square
                        time:   [10.713 ns 10.719 ns 10.726 ns]
                        change: [-14.165% -14.095% -14.026%] (p = 0.00 < 0.05)
field element operations/invert
                        time:   [4.0922 µs 4.0970 µs 4.1036 µs]
                        change: [-24.712% -24.549% -24.393%] (p = 0.00 < 0.05)
field element operations/sqrt
                        time:   [4.0430 µs 4.0470 µs 4.0525 µs]
                        change: [-24.461% -24.355% -24.224%] (p = 0.00 < 0.05)

     Running benches/scalar.rs
high-level operations/point-scalar mul
                        time:   [35.023 µs 35.036 µs 35.054 µs]
                        change: [-12.498% -12.413% -12.336%] (p = 0.00 < 0.05)
high-level operations/mul_by_generator naive
                        time:   [35.023 µs 35.031 µs 35.041 µs]
                        change: [-12.529% -12.444% -12.369%] (p = 0.00 < 0.05)
high-level operations/mul_by_generator precomputed
                        time:   [16.422 µs 16.425 µs 16.429 µs]
                        change: [-11.650% -11.576% -11.502%] (p = 0.00 < 0.05)
high-level operations/lincomb via mul+add
                        time:   [70.348 µs 70.637 µs 70.986 µs]
                        change: [-12.356% -12.186% -11.959%] (p = 0.00 < 0.05)
high-level operations/lincomb()
                        time:   [54.505 µs 54.580 µs 54.678 µs]
                        change: [-12.645% -12.500% -12.362%] (p = 0.00 < 0.05)
scalar operations/sub   
                        time:   [2.5533 ns 2.5546 ns 2.5564 ns]
                        change: [-26.584% -26.510% -26.441%] (p = 0.00 < 0.05)
scalar operations/add   
                        time:   [3.8494 ns 3.8515 ns 3.8541 ns]
                        change: [-22.241% -22.157% -22.077%] (p = 0.00 < 0.05)
scalar operations/mul   
                        time:   [27.433 ns 27.447 ns 27.466 ns]
                        change: [-1.2385% -1.0699% -0.9246%] (p = 0.00 < 0.05)
scalar operations/negate
                        time:   [4.7519 ns 4.7537 ns 4.7563 ns]
                        change: [+1.8517% +1.9516% +2.0527%] (p = 0.00 < 0.05)
scalar operations/invert
                        time:   [12.537 µs 12.545 µs 12.557 µs]
                        change: [-0.1999% -0.1031% -0.0059%] (p = 0.03 < 0.05)
```