@altkdf 2023-12-05
I likely overdid it with the `inline`s and some could probably be removed if that's important.