@tarcieri 2023-12-06
That's a lot of `inline(always)` specifically. We've had various reports of impacts on compile times in the past. I'm curious what the impact here would be.

I agree there are definitely measurable places it helps, but it would be nice to check each one's impact versus just an `#[inline]`