@altkdf 2023-12-06
> That's a lot of `inline(always)` specifically. We've had various reports of impacts on compile times in the past. I'm curious what the impact here would be.
> 
> I agree there are definitely measurable places it helps, but it would be nice to check each one's impact versus just an `#[inline]`

I agree. Would this be something that could be merged if I reduce the inlines to the minimum subset for the improvement? 

However, in general it *might* make sense to `inline` some calls up to the caller s.t. the compiler has as lest the theoretical possibility for some optimizations like vectorization or elimination of redundant computation. This could be the case if we e.g. do several multiplications/squarings in the same function. To decide if that's really the case, we would need additional benchmarks, of course. In case it turns out that we do need many `inline`s, maybe adding a feature flag for inlining would also be an option as well for performance-critical applications to reduce the compilation time for applications where it is not important?

without `inline`s
`RUSTFLAGS='-C target-cpu=native' cargo rebuild --release  20.68s user 8.13s system 444% cpu 6.478 total`
with `inline`s
`RUSTFLAGS='-C target-cpu=native' cargo rebuild --release  27.51s user 8.90s system 342% cpu 10.637 total`